middleware: restore time alloted for first-scan
JIRA: EDISON-1523

after wpa-supplicant bring-up, wpa_cli used to have 6 secs
to scan for networks. Due to a change in how the countdown
timer was implemented, this time got reduced to 3 secs. This
was not enough time for scanning right after bringing up
wpa_supplicant. Restored behavior to what it originally was.

Change-Id: Id5623fee8dbfc746efee69412dda80e7cf9981e7
Signed-off-by: Akshay Dua <akshay.dua@intel.com>
Reviewed-on: http://git-ger-5.devtools.intel.com/gerrit/124078
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Chereau, Fabien <fabien.chereau@intel.com>
Tested-by: Chereau, Fabien <fabien.chereau@intel.com>