middleware: Simplify WiFi setup in AP-mode
JIRA: EDISON-1305

allow user to pick SSID from a scanned list.
The scanning is done right before starting
AP-mode since a network scan is not possible
while in AP-mode (hardware limitation). The
user can also choose the manual setup option
for hidden networks and for those that did
not show up in the scanned list.

Change-Id: I345fb50c5651d05917f5b0fb6b702dbc5e5873b4
Signed-off-by: Akshay Dua <akshay.dua@intel.com>
Reviewed-on: https://git-ger-5.devtools.intel.com/gerrit/131098
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Chenthati, PradeepX <pradeepx.chenthati@intel.com>